feat: 통계 임시 데이터 구현
