--- presentation/src/main/java/com/hihihihi/gureumpage/ui/statistics/Genre.kt
@@ -1,13 +1,0 @@
-package com.hihihihi.gureumpage.ui.statistics
-
-data class Genre(
-    val value: Float,
-    val label: String,
-)
-
-val mockGenreList = listOf(
-    Genre(3f, "소설",),
-    Genre(4f, "판타지"),
-    Genre(5f, "자기계발"),
-)
-
